refactor(openai-bridge): split runtime retry layers
